fix: filter for webhook page now contains all fields
