+++ packages/frontend/app/shared/enums.ts
@@ -1,0 +1,8 @@
+export enum WebhookEventType {
+  IncomingPaymentCreated = 'incoming_payment.created',
+  IncomingPaymentCompleted = 'incoming_payment.completed',
+  IncomingPaymentExpired = 'incoming_payment.expired',
+  OutgoingPaymentCreated = 'outgoing_payment.created',
+  OutgoingPaymentCompleted = 'outgoing_payment.completed',
+  OutgoingPaymentFailed = 'outgoing_payment.failed'
+}
